# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CoronaMecca API",
        "description": "CoronaMecca API.",
        "contact": {
            "email": "dann4624@edu.sde.dk"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://vacapi.semeicardia.online/api"
        }
    ],
    "paths": {
...
        "/permissions": {
            "get": {
                "tags": [
                    "Permissions"
                ],
                "summary": "Get a list of permissions",
                "description": "Get a list of permissions",
                "operationId": "PermissionsList",
                "responses": {
                    "200": {
                        "description": "List of Permissions"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/permissions/{id}": {
            "get": {
                "tags": [
                    "Permissions"
                ],
                "summary": "Get a specific  an permission",
                "description": "Get a specific permission",
                "operationId": "PermissionsShow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the permission",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Permission object"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}